move retrieveMedicationInventory() to the proper service
